+++ dto/src/main/java/com/krillsson/sysapi/dto/storage/StorageInfo.java
@@ -14,11 +14,11 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class StorageInfo {
 
     @JsonProperty("openFileDescriptors")
-    private Integer openFileDescriptors;
+    private long openFileDescriptors;
     @JsonProperty("maxFileDescriptors")
-    private Integer maxFileDescriptors;
+    private long maxFileDescriptors;
     @JsonProperty("timeStamp")
-    private Integer timeStamp;
+    private long timeStamp;
     @JsonProperty("diskInfo")
     private DiskInfo[] diskInfo = null;
 
@@ -34,7 +34,7 @@ public class StorageInfo {
      * @param openFileDescriptors
      * @param maxFileDescriptors
      */
-    public StorageInfo(Integer openFileDescriptors, Integer maxFileDescriptors, Integer timeStamp, DiskInfo[] diskInfo) {
+    public StorageInfo(long openFileDescriptors, long maxFileDescriptors, long timeStamp, DiskInfo[] diskInfo) {
         super();
         this.openFileDescriptors = openFileDescriptors;
         this.maxFileDescriptors = maxFileDescriptors;
@@ -43,32 +43,32 @@ public class StorageInfo {
     }
 
     @JsonProperty("openFileDescriptors")
-    public Integer getOpenFileDescriptors() {
+    public long getOpenFileDescriptors() {
         return openFileDescriptors;
     }
 
     @JsonProperty("openFileDescriptors")
-    public void setOpenFileDescriptors(Integer openFileDescriptors) {
+    public void setOpenFileDescriptors(long openFileDescriptors) {
         this.openFileDescriptors = openFileDescriptors;
     }
 
     @JsonProperty("maxFileDescriptors")
-    public Integer getMaxFileDescriptors() {
+    public long getMaxFileDescriptors() {
         return maxFileDescriptors;
     }
 
     @JsonProperty("maxFileDescriptors")
-    public void setMaxFileDescriptors(Integer maxFileDescriptors) {
+    public void setMaxFileDescriptors(long maxFileDescriptors) {
         this.maxFileDescriptors = maxFileDescriptors;
     }
 
     @JsonProperty("timeStamp")
-    public Integer getTimeStamp() {
+    public long getTimeStamp() {
         return timeStamp;
     }
 
     @JsonProperty("timeStamp")
-    public void setTimeStamp(Integer timeStamp) {
+    public void setTimeStamp(long timeStamp) {
         this.timeStamp = timeStamp;
     }
